Move nfpm content yaml types
